# swagger.yaml
openapi: 3.0.0
info:
  title: Teamitaka Practice API
  version: 1.0.0
  description: |
    Teamitaka 프로젝트 연습을 위한 API 명세서로서, 사용자 관리, 인증, 프로젝트 관리, 모집 공고, 프로필, 댓글, 스크랩, 검색, 대학 인증 등의 기능을 제공합니다.
    이 API는 Node.js와 Express.js를 기반으로 구현되었으며, Swagger UI를 통해 시각화됩니다.  
    주요 기능:
    - 사용자 등록 및 인증 (JWT 기반)
    - 관리자 기능 (관리자 계정 관리, 인증된 이메일 관리)
    - 모집 공고 생성/수정/삭제 및 댓글 관리
    - 프로젝트 생성/수정/삭제 및 멤버 관리
    - 프로필 생성/수정/삭제 및 이력서 조회
    - 스크랩 기능
    - 키워드 기반 검색
    - 대학 인증 (Univecert API 연동)
  contact:
    name: Teamitaka Support
    email: support@teamitaka.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: 로컬 개발 서버
    variables:
      port:
        default: "3000"
        description: "서버 포트 (동적으로 설정 가능)"
  - url: https://api.teamitaka.com
    description: 프로덕션 서버
...
paths:
...
  /api/profiles/{user_id}:
    get:
      tags:
        - Profile
      summary: 사용자 프로필 조회
      description: 특정 사용자의 프로필 정보를 반환합니다.
      operationId: getProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 프로필을 조회할 사용자 ID
      responses:
        '200':
          description: 프로필 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  user_id: 1
                  nickname: "johndoe"
                  university: "고려대학교"
                  created_at: "2025-02-27T11:00:00.000Z"
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: ""
        '404':
          description: 프로필 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필을 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/profiles:
    post:
      tags:
        - Profile
      summary: 프로필 생성
      description: 새로운 프로필을 생성합니다.
      operationId: createProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                  example: 1
                  description: 프로필을 생성할 사용자 ID
                nickname:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "johndoe"
                  description: 사용자 닉네임 (1~50자)
                university:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "고려대학교"
                  description: 소속 대학 (1~100자)
              required:
                - user_id
                - nickname
                - university
      responses:
        '201':
          description: 프로필 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  user_id: 1
                  nickname: "johndoe"
                  university: "고려대학교"
                  created_at: "2025-02-27T11:00:00.000Z"
                message: "프로필 생성 완료"
        '400':
          description: 필수 필드 누락 또는 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "닉네임은 1~50자 이내여야 합니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '409':
          description: 이미 프로필이 존재함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "해당 사용자에 대한 프로필이 이미 존재합니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/profiles/{profile_id}:
    put:
      tags:
        - Profile
      summary: 프로필 수정
      description: 특정 프로필 정보를 수정합니다.
      operationId: updateProfile
      security:
        - bearerAuth: []
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
          description: 수정할 프로필 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "johndoe2"
                  description: 수정할 닉네임 (1~50자)
                university:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "서울대학교"
                  description: 수정할 소속 대학 (1~100자)
              required:
                - nickname
                - university
      responses:
        '200':
          description: 프로필 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  user_id: 1
                  nickname: "johndoe2"
                  university: "서울대학교"
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: "프로필 수정 완료"
        '400':
          description: 필수 필드 누락 또는 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "닉네임은 1~50자 이내여야 합니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필 수정 권한이 없습니다."
        '404':
          description: 프로필 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필을 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    delete:
      tags:
        - Profile
      summary: 프로필 삭제
      description: 특정 프로필을 삭제합니다.
      operationId: deleteProfile
      security:
        - bearerAuth: []
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: integer
          description: 삭제할 프로필 ID
      responses:
        '200':
          description: 프로필 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data: {}
                message: "프로필 삭제 완료"
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필 삭제 권한이 없습니다."
        '404':
          description: 프로필 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필을 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/profiles/resume/{user_id}:
    get:
      tags:
        - Profile
      summary: 사용자 이력서 조회
      description: 특정 사용자의 이력서 정보를 반환합니다. 이력서에는 기본 프로필 정보와 참여한 프로젝트 목록이 포함됩니다.
      operationId: getResume
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          description: 이력서를 조회할 사용자 ID
      responses:
        '200':
          description: 이력서 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  profile:
                    user_id: 1
                    nickname: "johndoe"
                    university: "고려대학교"
                  projects:
                    - project_id: 1
                      title: "프로젝트 A"
                      description: "프로젝트 설명"
                      status: "진행중"
                      role: "팀원"
                      joined_at: "2025-02-27T11:00:00.000Z"
                message: ""
        '404':
          description: 프로필 또는 이력서 정보 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "프로필을 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 댓글 관련 엔드포인트
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        data:
          type: object
          description: 반환 데이터
        message:
          type: string
          description: 추가 메시지
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: 요청 실패 여부
        message:
          type: string
          description: 오류 메시지
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 기반 인증 토큰 (Bearer 형식)
...